requirements.txt

streamlit==1.31.0
langchain==0.1.0
langchain-google-genai==0.0.7
google-generativeai==0.3.2
huggingface-hub==0.20.2
sentence-transformers==2.3.1
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.0
tiktoken==0.5.2